chore: split type adapters of adventure types from the main type adapter factory
