+++ data/json/src/main/java/net/hypejet/jet/data/json/DataTypeAdapterFactory.java
@@ -4,15 +4,11 @@ import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
 import com.google.gson.TypeAdapterFactory;
 import com.google.gson.reflect.TypeToken;
-import net.hypejet.jet.data.json.adapters.BinaryTagHolderTypeAdapter;
 import net.hypejet.jet.data.json.adapters.DataEntryTypeAdapter;
 import net.hypejet.jet.data.json.adapters.FeaturePackTypeAdapter;
-import net.hypejet.jet.data.json.adapters.KeyTypeAdapter;
 import net.hypejet.jet.data.json.adapters.UnitTypeAdapter;
 import net.hypejet.jet.data.json.entry.DataEntry;
 import net.hypejet.jet.data.json.util.JsonUnit;
-import net.kyori.adventure.key.Key;
-import net.kyori.adventure.nbt.api.BinaryTagHolder;
 import org.jspecify.annotations.NonNull;
 
 import java.lang.reflect.ParameterizedType;
@@ -40,12 +36,8 @@ public final class DataTypeAdapterFactory implements TypeAdapterFactory {
         Class<? super T> rawType = type.getRawType();
         Type actualType = type.getType();
 
-        if (Key.class.isAssignableFrom(rawType)) {
-            return KeyTypeAdapter.INSTANCE;
-        } else if (DataEntry.FeaturePack.class.isAssignableFrom(rawType)) {
+        if (DataEntry.FeaturePack.class.isAssignableFrom(rawType)) {
             return FeaturePackTypeAdapter.INSTANCE;
-        } else if (BinaryTagHolder.class.isAssignableFrom(rawType)) {
-            return BinaryTagHolderTypeAdapter.INSTANCE;
         } else if (JsonUnit.class.isAssignableFrom(rawType)) {
             return UnitTypeAdapter.INSTANCE;
         } else if (DataEntry.class.isAssignableFrom(rawType)) {
